requirements.txt
Markdown==3.7
openai==1.64.0
pydantic==2.10.6
python-dotenv==1.0.1
python-telegram-bot==21.10
weasyprint==64.1
